+++ Exercise_3/Exercise_3.3/Fibonacci_1.h
@@ -1,3 +1,4 @@
+#include <stdio.h>
 #pragma once
 
 /** @file Fibonacci_1.h
@@ -5,7 +6,7 @@
  */
 
 /** @brief Calculates the fibonacci number at the place of the given number
- *  @param the number which determines the place of the fibonacci number
+ *  @param num  the number which determines the place of the fibonacci number
  *  @returns the fibonacci number at the corresponding place
  */
-int GetFibNumber(int num);
+long long int GetFibNumber(size_t num);
